Revert "Merge branch 'main' into Game_Over"
This reverts commit fc7ee512039196c9592502232e652a72f533f3c2, reversing
changes made to 9e0186f08d6a2d0865b0b25fb0c5ab3f007ea74c.

+++ Assets/Scripts/EnemyController.cs
@@ -7,7 +7,8 @@ public class EnemyController : MonoBehaviour
     [SerializeField] private float fireRate;
     [SerializeField] private float bulletSpeedModifier;
     [SerializeField] private int startingHealth;
-    private float health;
+    
+    private int health;
     private PathController[] pathControllers;
     private Transform bulletSpawnPoint;
     private int currentMovementIdx = 0;
@@ -50,7 +51,7 @@ public class EnemyController : MonoBehaviour
         simpleBullet.SetSize(0.3f);
 
     }
-    public void Damage(float damage)
+    public void Damage(int damage)
     {
         health -= damage;
         if (health <= 0)
